Fix duration type from double to int in Cancion

+++ Cancion.java
@@ -2,7 +2,7 @@ import java.util.Scanner;
 public class Cancion extends ContMulti {
     private String artista;
 
-    public Cancion(String titulo, double duracion, String artista) {
+    public Cancion(String titulo, int duracion, String artista) {
         super(titulo, duracion);
         this.artista = artista;
     }
@@ -46,6 +46,7 @@ public class Cancion extends ContMulti {
 }
  
     
+
